fix: widen residual int→long for platform record IDs (Phase 5 Task 5.5)
Widen all missed id-typed `int` members to `long`:
- `LinkStatPoint.LinkId` int→long
- `CampaignData.Groups` IReadOnlyList<int>→IReadOnlyList<long>
- `CampaignGroupInfo.Id` int→long (CardsCnt left as int — it is a count)
- `MessageListCriteria.CampaignIds` IReadOnlyList<int>→IReadOnlyList<long>
- `MessageListCriteria.Groups` IReadOnlyList<int>→IReadOnlyList<long>
- `AlphanameData.Id` added as long (fixture field "id" was unmapped)
- `AlphanameData.AlphanameId` int→long
- `AlphanameInfo.Id` int→long (Type/GlobalStatus/PartnerStatus/IsDefault left as int)

Tests: add LinkId round-trip assertion in GetStatsAsync; assert AlphanameData.Id
and AlphanameInfo.Id from fixture (7719 and 58356 respectively). 171 tests pass.

+++ src/Mobizon.Contracts/Models/Campaigns/CampaignData.cs
@@ -90,7 +90,7 @@ namespace Mobizon.Contracts.Models.Campaigns
         public bool? TrackShortLinkRecipients { get; set; }
 
         /// <summary>Gets or sets the contact group IDs used in the campaign.</summary>
-        public IReadOnlyList<int>? Groups { get; set; }
+        public IReadOnlyList<long>? Groups { get; set; }
 
         /// <summary>Gets or sets the moderator's comment if the campaign was declined.</summary>
         public string? GlobalComment { get; set; }
@@ -102,7 +102,7 @@ namespace Mobizon.Contracts.Models.Campaigns
     public class CampaignGroupInfo
     {
         /// <summary>Gets or sets the group ID.</summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>Gets or sets the group name.</summary>
         public string? Name { get; set; }
